Remove map onboarding and update e2e tests
