+++ app/views/onboarding/Onboarding5.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import {
   Dimensions,
+  FlatList,
   ImageBackground,
   StatusBar,
   StyleSheet,
@@ -9,10 +10,10 @@ import {
 
 import BackgroundImage from './../../assets/images/launchScreenBackground.png';
 import ButtonWrapper from '../../components/ButtonWrapper';
-import { Typography } from '../../components/Typography';
+import { Type, Typography } from '../../components/Typography';
 import Colors from '../../constants/colors';
 import fontFamily from '../../constants/fonts';
-import { LOCATION_DATA } from '../../constants/storage';
+import { AUTHORITY_SOURCE_SETTINGS } from '../../constants/storage';
 import { GetStoreData } from '../../helpers/General';
 import languages from '../../locales/languages';
 
@@ -22,26 +23,49 @@ class Onboarding extends Component {
   constructor(props) {
     super(props);
 
-    let storedLocations = GetStoreData(LOCATION_DATA, false);
-    storedLocations = Array.isArray(storedLocations) ? storedLocations : [];
     this.state = {
-      locationData: storedLocations,
+      selectedAuthorities: [],
     };
   }
 
-  hasLocationData = () => {
-    return (
-      Array.isArray(this.state.locationData) &&
-      this.state.locationData.length > 0
-    );
+  componentDidMount() {
+    this.props.navigation.addListener('focus', this.getAuthorities);
+    this.getAuthorities();
+  }
+
+  getAuthorities = () => {
+    GetStoreData(AUTHORITY_SOURCE_SETTINGS, false).then(result => {
+      if (result !== null) {
+        console.log('Retrieving settings from async storage:');
+        console.log(result);
+        this.setState({
+          selectedAuthorities: result,
+        });
+      } else {
+        console.log('No stored authority settings.');
+      }
+    });
   };
 
   onGoToSettingsPress = () => {
-    this.props.navigation.navigate('ImportScreen');
+    this.props.navigation.navigate('ChooseProviderScreen');
   };
 
   onButtonPress = () => {
     this.props.navigation.replace('Onboarding6');
+  };
+
+  renderAuthorities = () => {
+    return (
+      <FlatList
+        data={this.state.selectedAuthorities}
+        renderItem={({ item }) => (
+          <View style={styles.flatlistRowView}>
+            <Typography style={styles.item}>{item.key}</Typography>
+          </View>
+        )}
+      />
+    );
   };
 
   render() {
@@ -54,33 +78,29 @@ class Onboarding extends Component {
         />
         <View style={styles.mainContainer}>
           <View style={styles.contentContainer}>
-            <Typography style={styles.headerText}>
-              {languages.t('label.maps_import_title')}
+            <Typography
+              testID='Header'
+              style={styles.headerText}
+              use={Type.Headline2}>
+              {languages.t('label.choose_provider_title')}
             </Typography>
-            <Typography style={styles.subheaderText}>
-              {languages.t('label.maps_import_text')}
+            <Typography testID='Subheader' style={styles.subheaderText}>
+              {languages.t('label.choose_provider_subtitle')}
             </Typography>
-            <View style={styles.statusContainer}>
-              <View style={styles.spacer} />
-            </View>
+            <View style={styles.spacer} />
+            <View style={styles.listContainer}>{this.renderAuthorities()}</View>
           </View>
           <View style={styles.footerContainer}>
-            {!this.hasLocationData() && (
-              <ButtonWrapper
-                title={languages.t('label.maps_import_button_text')}
-                onPress={this.onGoToSettingsPress}
-                buttonColor={Colors.VIOLET}
-                bgColor={Colors.WHITE}
-              />
-            )}
+            <ButtonWrapper
+              title={languages.t('label.choose_provider_title')}
+              onPress={this.onGoToSettingsPress}
+              buttonColor={Colors.VIOLET}
+              bgColor={Colors.WHITE}
+            />
             <View style={styles.spacer} />
             <ButtonWrapper
-              testID='NextBtn'
-              title={
-                this.hasLocationData()
-                  ? languages.t('label.launch_next')
-                  : 'Skip'
-              }
+              testID='NextButton'
+              title={languages.t('label.launch_next')}
               onPress={this.onButtonPress}
               buttonColor={Colors.VIOLET}
               bgColor={Colors.WHITE}
@@ -110,9 +130,7 @@ const styles = StyleSheet.create({
   },
   headerText: {
     color: Colors.WHITE,
-    fontSize: 26,
     width: width * 0.8,
-    fontFamily: fontFamily.primaryMedium,
   },
   subheaderText: {
     marginTop: '3%',
@@ -121,17 +139,33 @@ const styles = StyleSheet.create({
     width: width * 0.55,
     fontFamily: fontFamily.primaryRegular,
   },
-  statusContainer: {
-    marginTop: '5%',
-  },
   spacer: {
-    marginVertical: '5%',
+    marginVertical: 7.5,
   },
   footerContainer: {
     position: 'absolute',
     bottom: 0,
     marginBottom: '10%',
     alignSelf: 'center',
+  },
+  item: {
+    fontFamily: fontFamily.primaryRegular,
+    fontSize: 16,
+    padding: 10,
+    maxWidth: '90%',
+    color: Colors.WHITE,
+  },
+  flatlistRowView: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    paddingTop: 7,
+    paddingBottom: 5,
+    borderBottomWidth: 1,
+    borderColor: Colors.SILVER,
+    overflow: 'scroll',
+  },
+  listContainer: {
+    height: '25%',
   },
 });
