+++ app/views/onboarding/Onboarding6.js
@@ -1,83 +1,258 @@
 import React, { Component } from 'react';
 import {
   Dimensions,
-  FlatList,
   ImageBackground,
   StatusBar,
   StyleSheet,
   View,
 } from 'react-native';
+import {
+  PERMISSIONS,
+  RESULTS,
+  check,
+  checkNotifications,
+  request,
+  requestNotifications,
+} from 'react-native-permissions';
+import { SvgXml } from 'react-native-svg';
 
 import BackgroundImage from './../../assets/images/launchScreenBackground.png';
+import IconDenied from '../../assets/svgs/permissionDenied';
+import IconGranted from '../../assets/svgs/permissionGranted';
+import IconUnknown from '../../assets/svgs/permissionUnknown';
 import ButtonWrapper from '../../components/ButtonWrapper';
-import { Typography } from '../../components/Typography';
+import { Type, Typography } from '../../components/Typography';
 import Colors from '../../constants/colors';
 import fontFamily from '../../constants/fonts';
-import { AUTHORITY_SOURCE_SETTINGS } from '../../constants/storage';
-import { GetStoreData } from '../../helpers/General';
+import { PARTICIPATE } from '../../constants/storage';
+import { SetStoreData } from '../../helpers/General';
 import languages from '../../locales/languages';
+import { isPlatformiOS } from '../../Util';
 
 const width = Dimensions.get('window').width;
+
+const PermissionStatusEnum = {
+  UNKNOWN: 0,
+  GRANTED: 1,
+  DENIED: 2,
+};
+
+const PermissionDescription = ({ title, status }) => {
+  let icon;
+  switch (status) {
+    case PermissionStatusEnum.UNKNOWN:
+      icon = IconUnknown;
+      break;
+    case PermissionStatusEnum.GRANTED:
+      icon = IconGranted;
+      break;
+    case PermissionStatusEnum.DENIED:
+      icon = IconDenied;
+      break;
+  }
+  return (
+    <View style={styles.permissionContainer}>
+      <Typography style={styles.permissionTitle}>{title}</Typography>
+      <SvgXml style={styles.permissionIcon} xml={icon} width={30} height={30} />
+    </View>
+  );
+};
 
 class Onboarding extends Component {
   constructor(props) {
     super(props);
-
     this.state = {
-      selectedAuthorities: [],
+      notificationPermission: PermissionStatusEnum.UNKNOWN,
+      locationPermission: PermissionStatusEnum.UNKNOWN,
     };
+    this.checkLocationStatus();
+    this.checkNotificationStatus();
   }
 
-  componentDidMount() {
-    this.props.navigation.addListener('focus', this.getAuthorities);
-    this.getAuthorities();
+  isLocationChecked() {
+    return this.state.locationPermission !== PermissionStatusEnum.UNKNOWN;
   }
 
-  getAuthorities = () => {
-    GetStoreData(AUTHORITY_SOURCE_SETTINGS, false).then(result => {
-      if (result !== null) {
-        console.log('Retrieving settings from async storage:');
-        console.log(result);
-        this.setState({
-          selectedAuthorities: result,
-        });
-      } else {
-        console.log('No stored authority settings.');
+  isNotificationChecked() {
+    return this.state.notificationPermission !== PermissionStatusEnum.UNKNOWN;
+  }
+
+  checkLocationStatus() {
+    // NEED TO TEST ON ANNDROID
+    let locationPermission;
+    if (isPlatformiOS()) {
+      locationPermission = PERMISSIONS.IOS.LOCATION_ALWAYS;
+    } else {
+      locationPermission = PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION;
+    }
+    check(locationPermission)
+      .then(result => {
+        switch (result) {
+          case RESULTS.GRANTED:
+            this.setState({
+              locationPermission: PermissionStatusEnum.GRANTED,
+            });
+            break;
+          case RESULTS.UNAVAILABLE:
+          case RESULTS.BLOCKED:
+            this.setState({
+              locationPermission: PermissionStatusEnum.DENIED,
+            });
+            break;
+        }
+      })
+      .catch(error => {
+        console.log('error checking location: ' + error);
+      });
+  }
+
+  checkNotificationStatus() {
+    checkNotifications().then(({ status }) => {
+      switch (status) {
+        case RESULTS.GRANTED:
+          this.setState({
+            notificationPermission: PermissionStatusEnum.GRANTED,
+          });
+          break;
+        case RESULTS.UNAVAILABLE:
+        case RESULTS.BLOCKED:
+          this.setState({
+            notificationPermission: PermissionStatusEnum.DENIED,
+          });
+          break;
       }
     });
-  };
+  }
 
-  hasAuthorities = () => {
+  requestLocation() {
+    // NEED TO TEST ON ANNDROID
+    let locationPermission;
+    if (isPlatformiOS()) {
+      locationPermission = PERMISSIONS.IOS.LOCATION_ALWAYS;
+    } else {
+      locationPermission = PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION;
+    }
+    request(locationPermission).then(result => {
+      switch (result) {
+        case RESULTS.GRANTED:
+          console.log('Location granted');
+          this.setState({
+            locationPermission: PermissionStatusEnum.GRANTED,
+          });
+          break;
+        case RESULTS.UNAVAILABLE:
+        case RESULTS.BLOCKED:
+          this.setState({
+            locationPermission: PermissionStatusEnum.DENIED,
+          });
+          break;
+      }
+    });
+  }
+
+  requestNotification() {
+    requestNotifications(['alert', 'badge', 'sound']).then(({ status }) => {
+      switch (status) {
+        case RESULTS.GRANTED:
+          this.setState({
+            notificationPermission: PermissionStatusEnum.GRANTED,
+          });
+          break;
+        case RESULTS.UNAVAILABLE:
+        case RESULTS.BLOCKED:
+          this.setState({
+            notificationPermission: PermissionStatusEnum.DENIED,
+          });
+          break;
+      }
+    });
+  }
+
+  buttonPressed() {
+    if (!this.isLocationChecked()) {
+      this.requestLocation();
+    } else if (!this.isNotificationChecked()) {
+      this.requestNotification();
+    } else {
+      SetStoreData(PARTICIPATE, 'true'); // replaces "start" button
+      SetStoreData('ONBOARDING_DONE', true);
+      this.props.navigation.replace('LocationTrackingScreen');
+    }
+  }
+
+  getTitleText() {
+    if (!this.isLocationChecked()) {
+      return languages.t('label.launch_location_header');
+    } else if (!this.isNotificationChecked()) {
+      return languages.t('label.launch_notif_header');
+    } else {
+      return languages.t('label.launch_done_header');
+    }
+  }
+
+  getTitleTextView() {
+    if (!this.isLocationChecked() || !this.isNotificationChecked()) {
+      return (
+        <Typography style={styles.headerText} use={Type.Headline2}>
+          {this.getTitleText()}
+        </Typography>
+      );
+    } else {
+      return (
+        <Typography style={styles.bigHeaderText} use={Type.Headline1}>
+          {this.getTitleText()}
+        </Typography>
+      );
+    }
+  }
+
+  getSubtitleText() {
+    if (!this.isLocationChecked()) {
+      return languages.t('label.launch_location_subheader');
+    } else if (!this.isNotificationChecked()) {
+      return languages.t('label.launch_notif_subheader');
+    } else {
+      return languages.t('label.launch_done_subheader');
+    }
+  }
+
+  getLocationPermission() {
     return (
-      Array.isArray(this.state.selectedAuthorities) &&
-      this.state.selectedAuthorities.length > 0
+      <>
+        <View style={styles.divider} />
+        <PermissionDescription
+          title={languages.t('label.launch_location_access')}
+          status={this.state.locationPermission}
+        />
+        <View style={styles.divider} />
+      </>
     );
-  };
+  }
 
-  onGoToSettingsPress = () => {
-    this.props.navigation.navigate('ChooseProviderScreen');
-  };
+  getNotificationsPermissionIfIOS() {
+    if (isPlatformiOS()) {
+      return (
+        <>
+          <PermissionDescription
+            title={languages.t('label.launch_notification_access')}
+            status={this.state.notificationPermission}
+          />
+          <View style={styles.divider} />
+        </>
+      );
+    }
+    return;
+  }
 
-  onButtonPress = () => {
-    this.props.navigation.replace('Onboarding7');
-  };
-
-  renderAuthorities = () => {
-    return this.hasAuthorities() ? (
-      <FlatList
-        data={this.state.selectedAuthorities}
-        renderItem={({ item }) => (
-          <View style={styles.flatlistRowView}>
-            <Typography style={styles.item}>{item.key}</Typography>
-          </View>
-        )}
-      />
-    ) : (
-      <Typography style={styles.subheaderText}>
-        {languages.t('label.authorities_no_sources')}
-      </Typography>
-    );
-  };
+  getButtonText() {
+    if (!this.isLocationChecked()) {
+      return languages.t('label.launch_enable_location');
+    } else if (!this.isNotificationChecked()) {
+      return languages.t('label.launch_enable_notif');
+    } else {
+      return languages.t('label.launch_finish_set_up');
+    }
+  }
 
   render() {
     return (
@@ -87,34 +262,23 @@ class Onboarding extends Component {
           backgroundColor='transparent'
           translucent
         />
+
         <View style={styles.mainContainer}>
           <View style={styles.contentContainer}>
-            <Typography testID='Header' style={styles.headerText}>
-              {languages.t('label.choose_provider_title')}
+            {this.getTitleTextView()}
+            <Typography style={styles.subheaderText}>
+              {this.getSubtitleText()}
             </Typography>
-            <Typography testID='Subheader' style={styles.subheaderText}>
-              {languages.t('label.choose_provider_subtitle')}
-            </Typography>
-            <View style={styles.spacer} />
             <View style={styles.statusContainer}>
-              <Typography style={styles.subheaderText}>
-                {languages.t('label.authorities_title') + ':'}
-              </Typography>
-              {this.renderAuthorities()}
+              {this.getLocationPermission()}
+              {this.getNotificationsPermissionIfIOS()}
+              <View style={styles.spacer} />
             </View>
           </View>
           <View style={styles.footerContainer}>
             <ButtonWrapper
-              title={languages.t('label.choose_provider_title')}
-              onPress={this.onGoToSettingsPress}
-              buttonColor={Colors.VIOLET}
-              bgColor={Colors.WHITE}
-            />
-            <View style={styles.spacer} />
-            <ButtonWrapper
-              testID='NextButton'
-              title={languages.t('label.launch_next')}
-              onPress={this.onButtonPress}
+              title={this.getButtonText()}
+              onPress={this.buttonPressed.bind(this)}
               buttonColor={Colors.VIOLET}
               bgColor={Colors.WHITE}
             />
@@ -141,11 +305,15 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignSelf: 'center',
   },
+  bigHeaderText: {
+    color: Colors.WHITE,
+    lineHeight: 48.5,
+    paddingTop: 52 - 48.5, // lineHeight hack
+    width: width * 0.7,
+  },
   headerText: {
     color: Colors.WHITE,
-    fontSize: 26,
     width: width * 0.8,
-    fontFamily: fontFamily.primaryMedium,
   },
   subheaderText: {
     marginTop: '3%',
@@ -154,7 +322,14 @@ const styles = StyleSheet.create({
     width: width * 0.55,
     fontFamily: fontFamily.primaryRegular,
   },
-  statusContainer: {},
+  statusContainer: {
+    marginTop: '5%',
+  },
+  divider: {
+    backgroundColor: Colors.DIVIDER,
+    height: 1,
+    marginVertical: '3%',
+  },
   spacer: {
     marginVertical: '5%',
   },
@@ -164,12 +339,18 @@ const styles = StyleSheet.create({
     marginBottom: '10%',
     alignSelf: 'center',
   },
-  item: {
-    fontFamily: fontFamily.primaryRegular,
-    fontSize: 16,
-    padding: 10,
-    maxWidth: '90%',
+  permissionContainer: {
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+  },
+  permissionTitle: {
     color: Colors.WHITE,
+    fontSize: 16,
+    alignSelf: 'center',
+    fontFamily: fontFamily.primaryRegular,
+  },
+  permissionIcon: {
+    alignSelf: 'center',
   },
 });
