+++ e2e/Onboarding.spec.js
@@ -4,7 +4,6 @@ import Onboarding2 from './pages/Onboarding2.po.js';
 import Onboarding3 from './pages/Onboarding3.po.js';
 import Onboarding4 from './pages/Onboarding4.po.js';
 import Onboarding5 from './pages/Onboarding5.po.js';
-import Onboarding6 from './pages/Onboarding6.po.js';
 
 describe('Onboarding visual appearance', () => {
   it('Navigates through the onboarding without visual regression', async () => {
@@ -31,10 +30,6 @@ describe('Onboarding visual appearance', () => {
     await Onboarding5.isOnScreen();
     await Onboarding5.takeScreenshot();
     await Onboarding5.tapButton();
-
-    await Onboarding6.isOnScreen();
-    await Onboarding6.takeScreenshot();
-    await Onboarding6.tapButton();
 
     await EnableLocation.takeScreenshot();
     await EnableLocation.tapButton();
