+++ e2e/helpers/onboarding.js
@@ -3,7 +3,6 @@ import Onboarding2 from '../pages/Onboarding2.po.js';
 import Onboarding3 from '../pages/Onboarding3.po.js';
 import Onboarding4 from '../pages/Onboarding4.po.js';
 import Onboarding5 from '../pages/Onboarding5.po.js';
-import Onboarding6 from '../pages/Onboarding6.po.js';
 
 export const navigateThroughOnboarding = async () => {
   await Onboarding1.isOnScreen();
@@ -20,7 +19,4 @@ export const navigateThroughOnboarding = async () => {
 
   await Onboarding5.isOnScreen();
   await Onboarding5.tapButton();
-
-  await Onboarding6.isOnScreen();
-  await Onboarding6.tapButton();
 };
