+++ e2e/pages/Onboarding5.po.js
@@ -1,8 +1,7 @@
 /* eslint-disable */
-const buttonLabel = 'Skip';
-const screenTitle = 'Google Maps';
-const screenSubtitle =
-  'To see if you encountered someone with COVID-19 prior to downloading this app, you can import your personal location history.';
+const buttonLabel = 'Next';
+const screenTitle = 'Header';
+const screenSubtitle = 'Subheader';
 const screenshotText = 'Onboarding - Page 5';
 
 class Onboarding5 {
@@ -16,8 +15,8 @@ class Onboarding5 {
 
   async isOnScreen() {
     // eslint-disable-next-line jest/no-standalone-expect
-    await expect(element(by.text(screenTitle))).toBeVisible();
-    await expect(element(by.text(screenSubtitle))).toBeVisible();
+    await expect(element(by.id(screenTitle))).toBeVisible();
+    await expect(element(by.id(screenSubtitle))).toBeVisible();
   }
 }
